requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# App & API
streamlit
fastapi
uvicorn
pydantic
requests
google-genai

# Data & DB
pandas
numpy
python-dotenv
snowflake-connector-python

# Viz
altair

# RAG / ML
pypdf
rank-bm25
sentence-transformers
scikit-learn
torch
